First Commit for this game
It is a Planet destroy and Merging game, where you have to merge objects and it will create a new one, mergeable objects will automatically shoot the Planet or you can also shoot the Planet by tapping on the screen. As you merge further the attack of the mergeable objects also increases.