Fifth Commit for this game
Added 5x Damage Multiplier button and showed Post delay for both boosts on the HUD and also Fixed Grid view UI, afterward it has a problem with spawning New-Objects but it has been fixed.

Next: Change Mesh Objects for Mergeable Objects and add an animation Object which will give life to the Objects.