Sixth Commit for this game
Changed Mesh for the Mergeable Objects and setup animation for them (Animation for Throw and Auto-Click) and also made the Script to Instantiate only one Mergeable Prefab object and change their color when merged to differentiate.

Next: Make the Color to allocate in an array randomly by changing RGBA values.