Eighth Commit for this game
Added a Upgrade Panel and the button which its open and modified the Pause Panel and also fixed the Touchable and Un-Touchable GameObject Size with Screen, also added a footer Panel where Button to open Upgrade Panel and Buy button is Present, also added bonus Coins when destroy a Planet and added animation which gives the indicator for this. Did fix Some Bug Miner Bugs that were ignored earlier.

Next: Add Functionalities for Upgrades.