Fix: Projectile did not spawn when AutoClick is on
Fix AutoClick Projectile Spawn Problem and also made the mergeable objects set at any tile on the grid. Also Earlier I Forgot to commit, I added the Damage text with its animation that tells how much damage it inflicted on the player.

Next: The previous commit's task is not completed so I will do it in next commit but additionally, Add Text for Planet Health that tells how much health of the player is remaining and also increase the mergeable index after a certain level of merges.