Final Commit for this game
All the desired features of this game have been added and is now eligible for further publication.

In future, their will be updates on this game based on some bug fixes or adding minor or major features to the game.

I am Khawaja Ibrahim Salim sole Developer of "Planet Invasion".